refactor heal server code
