+++ pkg/networkservice/common/heal/gen.go
@@ -19,22 +19,12 @@
 package heal
 
 import (
-	"context"
 	"sync"
-
-	"github.com/networkservicemesh/api/pkg/api/networkservice"
 )
 
 //go:generate go-syncmap -output connection_info_map.gen.go -type connectionInfoMap<string,*connectionInfo>
 
 type connectionInfoMap sync.Map
-
-type ctxWrapper struct {
-	mut     sync.Mutex
-	request *networkservice.NetworkServiceRequest
-	ctx     context.Context
-	cancel  func()
-}
 
 //go:generate go-syncmap -output context_wrapper_map.gen.go -type ctxWrapperMap<string,*ctxWrapper>
